Use pytz to localize datetimes.
The database and naïve datetimes should be UTC only, but for display purposes we will want to have localizations.

We know exactly when the conference is, so DST shouldn't bite us. But it's the Right thing to do. And if we re-use it next year, we might be on the other side of DST (we nearly straddled it this year!).